extract source and sink creation
